Address PR review: fix API ordering, test isolation, and annotation ID bug
- Move SessionReleaseCapturedGraph to end of OrtApi struct (version 27 addition)
- Move ReleaseCapturedGraph to end of OrtEp struct, update version check to < 27
- Fix m_current_graph_annotation_id ordering: set before RefreshBufferManager()
  and IsGraphCaptureAllowed() read it
- Extract graph capture test to graph_capture_test.cc with multi-ID coverage:
  capture/replay two annotation IDs independently, verify cross-ID isolation,
  selective release, and re-capture after release

Tested:
- onnxruntime_provider_test --gtest_filter=*ReleaseCapturedGraph*: PASS
- onnxruntime_provider_test --gtest_filter=*TestGraphCapture*: PASS
- Multi-generator e2e with phi4-graph model: PASS
- verify_single.py stress test (M=30-128): PASS
