+++ onnxruntime/core/session/onnxruntime_c_api.cc
@@ -4922,10 +4922,10 @@ static constexpr OrtApi ort_api_1_to_27 = {
     &OrtApis::KernelInfoGetAttributeArray_string,
     &OrtApis::SetPerSessionThreadPoolCallbacks,
     // End of Version 25 - DO NOT MODIFY ABOVE (see above text for more information)
-    &OrtApis::SessionReleaseCapturedGraph,
     // End of Version 26 - DO NOT MODIFY ABOVE (see above text for more information)
     &OrtApis::GetMemPatternEnabled,
     &OrtApis::GetSessionExecutionMode,
+    &OrtApis::SessionReleaseCapturedGraph,
 };
 
 // OrtApiBase can never change as there is no way to know what version of OrtApiBase is returned by OrtGetApiBase.
@@ -4964,7 +4964,8 @@ static_assert(offsetof(OrtApi, GetEpApi) / sizeof(void*) == 317, "Size of versio
 static_assert(offsetof(OrtApi, CreateExternalInitializerInfo) / sizeof(void*) == 389, "Size of version 23 API cannot change");
 static_assert(offsetof(OrtApi, GetTensorElementTypeAndShapeDataReference) / sizeof(void*) == 414, "Size of version 24 API cannot change");
 static_assert(offsetof(OrtApi, SetPerSessionThreadPoolCallbacks) / sizeof(void*) == 418, "Size of version 25 API cannot change");
-static_assert(offsetof(OrtApi, SessionReleaseCapturedGraph) / sizeof(void*) == 419, "Size of version 26 API cannot change");
+// no additions in version 26
+static_assert(offsetof(OrtApi, SessionReleaseCapturedGraph) / sizeof(void*) == 421, "Size of version 27 API cannot change");
 
 // So that nobody forgets to finish an API version, this check will serve as a reminder:
 static_assert(std::string_view(ORT_VERSION) == "1.27.0",
