+++ onnxruntime/core/session/plugin_ep/ep_plugin_provider_interfaces.cc
@@ -1038,7 +1038,7 @@ Status PluginExecutionProvider::ReleaseCapturedGraph(int graph_annotation_id) {
   // For plugin EPs that don't implement ReleaseCapturedGraph (version < 26 or null function pointer),
   // fall back to the base class no-op implementation. This is intentional: the request is silently
   // ignored since the plugin EP doesn't support explicit graph resource release.
-  if (ort_ep_->ort_version_supported < 26 || ort_ep_->ReleaseCapturedGraph == nullptr) {
+  if (ort_ep_->ort_version_supported < 27 || ort_ep_->ReleaseCapturedGraph == nullptr) {
     return Base::ReleaseCapturedGraph(graph_annotation_id);
   }
   return ToStatusAndRelease(ort_ep_->ReleaseCapturedGraph(ort_ep_.get(), graph_annotation_id));
